writing selenium tests and error fixing

+++ starter/cloudstorage/src/main/java/com/udacity/jwdnd/course1/cloudstorage/controllers/CrudController.java
@@ -50,8 +50,12 @@ public class CrudController {
 	
 
 	@RequestMapping("/addNote.do")
-	public String noteCreate(@ModelAttribute("SpringWeb") Notes note, HttpSession session, Users user, Model model) {
+	public String noteCreate(@ModelAttribute("SpringWeb") Notes note, HttpSession session, Users user, Model model, HttpServletResponse response) {
 		user = (Users) session.getAttribute("user");
+		if(user == null) {
+			response.setStatus(403);
+			return "error";
+		}
 		if (note.getNoteid() != null) {
 			crudService.editNotes(note);
 			user.setNotes(notesMapper.findNoteByUserId(user.getUserid()));
@@ -68,8 +72,12 @@ public class CrudController {
 	}
 
 	@RequestMapping("/addNote.do{noteid}")
-	public String noteEdit(@RequestParam("note.noteid") Integer noteid, HttpSession session, Users user, Model model) {
+	public String noteEdit(@RequestParam("note.noteid") Integer noteid, HttpSession session, Users user, Model model,HttpServletResponse response) {
 		user = (Users) session.getAttribute("user");
+		if(user == null) {
+			response.setStatus(403);
+			return "error";
+		}
 		Notes note = notesMapper.findNoteById(noteid);
 		notesMapper.addNotetoDatabase(note, user.getUserid());
 		session.setAttribute("user", user);
@@ -78,8 +86,13 @@ public class CrudController {
 	}
 
 	@RequestMapping("/deleteNote.do{noteid}")
-	public String deleteNote(@RequestParam("noteid") Integer noteid, HttpSession session) {
+	public String deleteNote(@RequestParam("noteid") Integer noteid, HttpSession session,HttpServletResponse response) {
+		
 		Users user = (Users) session.getAttribute("user");
+		if(user == null) {
+			response.setStatus(403);
+			return "error";
+		}
 		notesMapper.deleteNote(noteid);
 		user.setNotes(notesMapper.findNoteByUserId(user.getUserid()));
 		session.setAttribute("user", user);
@@ -90,8 +103,15 @@ public class CrudController {
 	// C R E D E N T I A L S
 	@RequestMapping("/addCredential.do")
 	public String credentialsCreate(@ModelAttribute("SpringWeb") Credentials credential, HttpSession session,
-			Users user, Model model) {
+			Users user, Model model,HttpServletResponse response) {
 		user = (Users) session.getAttribute("user");
+		if(user == null) {
+			response.setStatus(403);
+			return "error";
+		}if(credential.getUserid() != user.getUserid()) {
+			response.setStatus(403);
+			return "error";
+		}
 		credential.setSkeleton(crudService.generateRandomKey());
 		String password = credential.getPassword();
 		System.out.println("********" + password);
@@ -114,11 +134,20 @@ public class CrudController {
 	}
 
 	@RequestMapping("/deleteCredential.do{credentialid}")
-	public String deleteCredential(@RequestParam("credentialid") Integer credentialid, HttpSession session) {
+	public String deleteCredential(@RequestParam("credentialid") Integer credentialid, HttpSession session,HttpServletResponse response) {
 		Users user = (Users) session.getAttribute("user");
+		if(user == null) {
+			response.setStatus(403);
+			return "error";
+		}
+		if(credentialsMapper.findCredentialById(credentialid).getUserid() != user.getUserid()) {
+			response.setStatus(403);
+			return "error";
+		}
 		credentialsMapper.deleteCredential(credentialid);
 		user.setCredentials(credentialsMapper.findCredentialByUserId(user.getUserid()));
 		session.setAttribute("user", user);
+		
 
 		return "home";
 	}
@@ -134,8 +163,12 @@ public class CrudController {
 	// F I L E S
 	
 	@RequestMapping("/addfile.do")
-	public String createFile(@ModelAttribute("SpringWeb") MultipartFile fileUpload, HttpSession session, Users user, Model model) {
+	public String createFile(@ModelAttribute("SpringWeb") MultipartFile fileUpload, HttpSession session, Users user, Model model,HttpServletResponse response) {
 		user = (Users) session.getAttribute("user");
+		if(user == null) {
+			response.setStatus(403);
+			return "error";
+		}
 
 		Files file = new Files();
 		try {
@@ -152,8 +185,28 @@ public class CrudController {
 		return "home";
 		
 	}
+	@RequestMapping("/deleteFile.do{fileid}")
+	public String deleteFile(@RequestParam("fileid") Integer fileid, HttpSession session,HttpServletResponse response) {
+		Users user = (Users) session.getAttribute("user");
+		if(user == null) {
+			response.setStatus(403);
+			return "error";
+		}
+		if(filesMapper.findFilesByFileId(fileid).getUserid() != user.getUserid()) {
+			response.setStatus(403);
+			return "error";
+		}
+		filesMapper.deleteFileFromDatabase(fileid);
+		user.setFiles(filesMapper.findFilesByUserId(user.getUserid()));
+		session.setAttribute("user", user);
+		
+
+		return "home";
+	}
+	
 	@GetMapping("/img/{id}")
 	public void getImage(@PathVariable("id") Integer id, HttpServletResponse response) {
+		
 	        crudService.writeImageToRespose(id, response);
 	}
 }
